Create an argument parser class
parse yaml
reconfigure some code
