+++ crawler.js
@@ -13,15 +13,7 @@ const warcio = require("warcio");
 
 const Redis = require("ioredis");
 
-const TextExtract = require("./textextract");
 const behaviors = fs.readFileSync("/app/node_modules/browsertrix-behaviors/dist/behaviors.js", "utf-8");
-
-const HTML_TYPES = ["text/html", "application/xhtml", "application/xhtml+xml"];
-const WAIT_UNTIL_OPTS = ["load", "domcontentloaded", "networkidle0", "networkidle2"];
-
-const BEHAVIOR_LOG_FUNC = "__bx_log";
-
-const CHROME_PATH = "google-chrome";
 
 // to ignore HTTPS error for HEAD check
 const HTTPS_AGENT = require("https").Agent({
@@ -30,14 +22,19 @@ const HTTPS_AGENT = require("https").Agent({
 
 const HTTP_AGENT = require("http").Agent();
 
-const { ScreenCaster, NewWindowPage } = require("./screencaster");
-
+const { TextExtract } = require("./util/textextract");
+const { ScreenCaster } = require("./util/screencaster");
+const { argParser } = require("./util/argParser");
+const { constants } = require("./util/constants");
 
 // ============================================================================
 class Crawler {
   constructor() {
     this.headers = {};
-
+    
+    this.argParser = new argParser();
+    this.constants = new constants();
+     
     this.seenList = new Set();
 
     this.emulateDevice = null;
@@ -54,16 +51,18 @@ class Crawler {
     this.userAgent = "";
     this.behaviorsLogDebug = false;
     this.profileDir = fs.mkdtempSync(path.join(os.tmpdir(), "profile-"));
-
-    const params = require("yargs")
+    
+    const params = fs.existsSync("/app/browsertrixArgsConfig.yaml") ? this.argParser.parseYaml() : require("yargs")
       .usage("crawler [options]")
       .option(this.cliOpts)
-      .check((argv) => this.validateArgs(argv)).argv;
-
-    console.log("Exclusions Regexes: ", params.exclude);
-    console.log("Scope Regexes: ", params.scope);
+      .check((argv) => this.argParser.validateArgs(argv)).argv;
 
     this.params = params;
+  
+    console.log("Exclusions Regexes: ", this.params.exclude);
+    console.log("Scope Regexes: ", this.params.scope);
+
+
     this.capturePrefix = `http://${process.env.PROXY_HOST}:${process.env.PROXY_PORT}/${this.params.collection}/record/id_/`;
 
     this.gotoOpts = {
@@ -329,173 +328,6 @@ class Crawler {
     };
   }
 
-  validateUserUrl(url) {
-    url = new URL(url);
-
-    if (url.protocol !== "http:" && url.protocol != "https:") {
-      throw new Error("URL must start with http:// or https://");
-    }
-
-    return url;
-  }
-
-  validateArgs(argv) {
-    let purl;
-
-    if (argv.url) {
-      // Scope for crawl, default to the domain of the URL
-      // ensure valid url is used (adds trailing slash if missing)
-      //argv.seeds = [Crawler.validateUserUrl(argv.url)];
-      purl = this.validateUserUrl(argv.url);
-      argv.url = purl.href;
-    }
-    
-    if (argv.url && argv.urlFile) {
-      console.warn("You've passed a urlFile param, only urls listed in that file will be processed. If you also passed a url to the --url flag that will be ignored.");
-    }
-
-    // Check that the collection name is valid.
-    if (argv.collection.search(/^[\w][\w-]*$/) === -1){
-      throw new Error(`\n${argv.collection} is an invalid collection name. Please supply a collection name only using alphanumeric characters and the following characters [_ - ]\n`);
-    }
-  
-    argv.timeout *= 1000;
-
-    // waitUntil condition must be: load, domcontentloaded, networkidle0, networkidle2
-    // can be multiple separate by comma
-    // (see: https://github.com/puppeteer/puppeteer/blob/main/docs/api.md#pagegotourl-options)
-    argv.waitUntil = argv.waitUntil.split(",");
-
-    for (const opt of argv.waitUntil) {
-      if (!WAIT_UNTIL_OPTS.includes(opt)) {
-        throw new Error("Invalid waitUntil option, must be one of: " + WAIT_UNTIL_OPTS.join(","));
-      }
-    }
-
-    // log options
-    argv.logging = argv.logging.split(",");
-
-    // background behaviors to apply
-    const behaviorOpts = {};
-    argv.behaviors.split(",").forEach((x) => behaviorOpts[x] = true);
-    if (argv.logging.includes("behaviors")) {
-      behaviorOpts.log = BEHAVIOR_LOG_FUNC;
-    } else if (argv.logging.includes("behaviors-debug")) {
-      behaviorOpts.log = BEHAVIOR_LOG_FUNC;
-      this.behaviorsLogDebug = true;
-    }
-    this.behaviorOpts = JSON.stringify(behaviorOpts);
-
-    if (!argv.newContext) {
-      argv.newContext = "page";
-    }
-
-    switch (argv.newContext) {
-    case "page":
-      argv.newContext = Cluster.CONCURRENCY_PAGE;
-      if (argv.screencastPort && argv.workers > 1) {
-        console.warn("Note: Screencast with >1 workers and default page context may only show one page at a time. To fix, add '--newContext window' to open each page in a new window");
-      }
-      break;
-
-    case "session":
-      argv.newContext = Cluster.CONCURRENCY_CONTEXT;
-      break;
-
-    case "browser":
-      argv.newContext = Cluster.CONCURRENCY_BROWSER;
-      break;
-
-    case "window":
-      argv.newContext = NewWindowPage;
-      break;
-
-    default:
-      throw new Error("Invalid newContext, must be one of: page, session, browser");
-    }
-
-    if (argv.mobileDevice) {
-      this.emulateDevice = puppeteer.devices[argv.mobileDevice];
-      if (!this.emulateDevice) {
-        throw new Error("Unknown device: " + argv.mobileDevice);
-      }
-    }
-
-    if (argv.useSitemap === true) {
-      const url = new URL(argv.url);
-      url.pathname = "/sitemap.xml";
-      argv.useSitemap = url.href;
-    }
-
-    // Support one or multiple exclude
-    if (argv.exclude) {
-      if (typeof(argv.exclude) === "string") {
-        argv.exclude = [new RegExp(argv.exclude)];
-      } else {
-        argv.exclude = argv.exclude.map(e => new RegExp(e));
-      }
-    } else {
-      argv.exclude = [];
-    }
-
-    // warn if both scope and scopeType are set
-    if (argv.scope && argv.scopeType) {
-      console.warn("You've specified a --scopeType and a --scope regex. The custom scope regex will take precedence, overriding the scopeType");
-    }
-
-    // Support one or multiple scopes set directly, or via scopeType
-    if (argv.scope) {
-      if (typeof(argv.scope) === "string") {
-        argv.scope = [new RegExp(argv.scope)];
-      } else {
-        argv.scope = argv.scope.map(e => new RegExp(e));
-      }
-    } else {
-
-      // Set scope via scopeType
-      if (!argv.scopeType) {
-        argv.scopeType = argv.urlFile ? "any" : "prefix";
-      }
-
-      if (argv.scopeType && argv.url) {
-        switch (argv.scopeType) {
-        case "page":
-          // allow scheme-agnostic URLS as likely redirects
-          argv.scope = [new RegExp("^" + this.rxEscape(argv.url).replace(purl.protocol, "https?:") + "#.+")];
-          argv.allowHashUrls = true;
-          break;
-
-        case "prefix":
-          argv.scope = [new RegExp("^" + this.rxEscape(argv.url.slice(0, argv.url.lastIndexOf("/") + 1)))];
-          break;
-
-        case "domain":
-          argv.scope = [new RegExp("^" + this.rxEscape(purl.origin + "/"))];
-          break;
-
-        case "any":
-          argv.scope = [];
-          break;
-
-        default:
-          throw new Error(`Invalid scope type "${argv.scopeType}" specified, valid types are: page, prefix, domain`);
-
-
-        }
-      }
-    }
-
-    // Resolve statsFilename
-    if (argv.statsFilename) {
-      argv.statsFilename = path.resolve(argv.cwd, argv.statsFilename);
-    }
-
-    if (argv.profile) {
-      child_process.execSync("tar xvfz " + argv.profile, {cwd: this.profileDir});
-    }
-
-    return true;
-  }
 
   get chromeArgs() {
     // Chrome Flags, including proxy server
@@ -515,7 +347,7 @@ class Crawler {
     // Puppeter Options
     return {
       headless: this.params.headless,
-      executablePath: CHROME_PATH,
+      executablePath: this.constants.CHROME_PATH,
       ignoreHTTPSErrors: true,
       args: this.chromeArgs,
       userDataDir: this.profileDir,
@@ -561,7 +393,7 @@ class Crawler {
       }
 
       if (this.behaviorOpts && !page.__bx_inited) {
-        await page.exposeFunction(BEHAVIOR_LOG_FUNC, (logdata) => this._behaviorLog(logdata));
+        await page.exposeFunction(this.constants.BEHAVIOR_LOG_FUNC, (logdata) => this._behaviorLog(logdata));
         await page.evaluateOnNewDocument(behaviors + `;\nself.__bx_behaviors.init(${this.behaviorOpts});`);
         page.__bx_inited = true;
       }
@@ -907,7 +739,7 @@ class Crawler {
 
       const mime = contentType.split(";")[0];
 
-      if (HTML_TYPES.includes(mime)) {
+      if (this.constants.HTML_TYPES.includes(mime)) {
         return true;
       }
 
@@ -946,10 +778,6 @@ class Crawler {
 
   sleep(time) {
     return new Promise(resolve => setTimeout(resolve, time));
-  }
-
-  rxEscape(string) {
-    return string.replace(/[-/\\^$*+?.()|[\]{}]/g, "\\$&");
   }
 
   async parseSitemap(url) {
