+++ util/textextract.js
